refactor: 행사 상세 페이지 컴포넌트 이동
