+++ src/app/event/[eventId]/components/event-form/EventForm.tsx
@@ -1,8 +1,8 @@
 'use client';
 
-import HubButton from '../HubButton';
-import DateButton from '../DateButton';
-import BottomBar from '../BottomBar';
+import HubButton from './components/HubButton';
+import DateButton from './components/DateButton';
+import BottomBar from './components/BottomBar';
 import useBottomSheet from '@/hooks/useBottomSheet';
 import { useMemo, useState } from 'react';
 import BottomSheet from '@/components/bottom-sheet/BottomSheet';
