+++ src/app/event/[eventId]/components/steps/common-steps/CommonSidoStep.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import { BIG_REGIONS, BigRegionsType } from '@/constants/regions';
-import SidoButton from '../../SidoButton';
+import SidoButton from '../components/SidoButton';
 import { useAtomValue } from 'jotai';
 import { useFormContext } from 'react-hook-form';
 import { EventFormValues } from '../../../form.type';
